requirements.txt
pydantic~=2.6.1
fastapi~=0.109.2
python-dotenv~=1.0.1
psycopg2-binary==2.9.10